Replace shared_ptr with unique_ptr.
shared_ptrs are extremely inefficient due to use of atomic reference counters and extra overhead for control block maintenance. unique_ptr is practically the same as using raw pointer bug with auto-free capability.
Note, that `make_unique` (or `make_shared` for the previous version) should be used instead of `new` to make sure the memory is freed even in case of raised exception during construction, but I couldn't find the C++ standard milvus uses, so I didn't know if `make_unique` is allowed.