test: [skip e2e] optimize get_image_tag_by_short_name function
Refactor the  function to improve performance and readability. Instead of making API requests to Docker Hub, the function now retrieves tags from the Harbor registry. It also filters the tags based on the provided architecture and selects the latest tag that matches the prefix. This change enhances the efficiency of retrieving image tags by short name.

Signed-off-by: zhuwenxing <wenxing.zhu@zilliz.com>