enhance: Avoid execute load segment task twice.
After the executor finishes executing a task, there can be a maximum interval of 500ms
before the QC observes the new leader view.During this 500ms interval, the task won't
be flagged as finished by the scheduler. However, the scheduler might resubmit this task,
which remains in the "started" state, as soon as the QC detects the new leader view
while the executor is in the process of executing the task,the task will be marked
as "finished" triggering the scheduler to cancel it.

This PR block load segment task's finish until qc saw the new leader
view, to avoid execute load segment task twice.

Signed-off-by: Wei Liu <wei.liu@zilliz.com>