enhance: Mark query node as read only after suspend
after querynode has been suspended, it's not allow to load
segment/channel on it, which means the node is read only. to be
compatible with resource group design, after query node has been
suspend, we remove it from it's original resource group, make it a read
only query node in replica. then two things will happens:
1. it's original resource group will be lacking of query nodes, query
   coord will assign new node to it.
2. querycoord will try to move out all segments/channels after querynode
   has been suspended

Signed-off-by: Wei Liu <wei.liu@zilliz.com>