enhance: Remove implicit dependencies between tasks in querycoord
When the RPC corresponding to a task returns successfully, the data
distribution should have reached the expected state. If the expected state
is not achieved, it indicates an unexpected issue, such as a node going offline.
In such cases, if the task still waits for the data distribution to meet
the expectation before completing, it relies on other tasks to correct
the distribution, leading to implicit dependencies between tasks.

This PR removes the requirement for tasks to wait for the expected data
distribution as a completion condition, thereby eliminating the implicit
dependencies between tasks.

Signed-off-by: Wei Liu <wei.liu@zilliz.com>