feat: [2.5]Add support for modifying max capacity of array fields (#41406)
feat: Add support for modifying max capacity of array fields

This commit adds support for modifying the max capacity of array fields
in the alterCollectionFieldTask function. It checks if the field is an
array type and then validates and updates the max capacity value. This
change improves the flexibility of array fields in the collection.

Issue: https://github.com/milvus-io/milvus/issues/41363
pr:https://github.com/milvus-io/milvus/pull/41404

---------

Signed-off-by: Xianhui.Lin <xianhui.lin@zilliz.com>