fix: make results of hybrid search consistent regardless of output fields
issue: https://github.com/milvus-io/milvus-sdk-java/issues/1450

We found that the TopK results of a hybrid search can differ when the result set is empty and different output fields are requested.
This issue is caused by the merging logic when no records pass the filter, resulting in inconsistent handling of empty result sets.

Signed-off-by: sunby <sunbingyi1992@gmail.com>