enhance: Introduce sparse filter in query
The current commit implements sparse filtering in query tasks
using the statistical information (Bloom filter/MinMax) of the
Primary Key (PK).

The statistical information of the PK is bound to the segment
during the segment loading phase. A new filter has been added
to the segment filter to enable the sparse filtering functionality.

Signed-off-by: jiaqizho <jiaqi.zhou@zilliz.com>