fix: append operation can be only canceled by the wal itself but not the rpc
We need to promise the state of wal kept consistent with the memory state of streamingnode.
So we don't allow the append operation can be cancelled by the append caller to avoid leave a inconsistent state of alive wal.
The wal append operation can only be cancelled when the wal is shutting down.

Signed-off-by: chyezh <chyezh@outlook.com>