remove the streamingnode checking when loading segment
If we enable checking when loading segments, all segment should always be loaded by streamingnode but not 2.5 querynode, make some search and query failure when upgrading.
Otherwise, some search and query result will be wrong when upgrading. We choose to disable this checking for now to promise available search and query when upgrading.

Signed-off-by: chyezh <chyezh@outlook.com>